+++ src/layouts/PublicLayout.jsx
@@ -1,7 +1,7 @@
 
 import { Outlet } from "react-router-dom";
-import Footer from "../componemts/Footer/Footer";
-import Navbar from "../componemts/Navbar/Navbar";
+import Footer from "../componemts/Layout/Footer/Footer";
+import Navbar from "../componemts/Layout/Navbar/Navbar";
 export const Layout = () => {
     return (
       <div className="app">
