+++ src/router.js
@@ -1,13 +1,13 @@
 import { createBrowserRouter } from "react-router-dom";
-import About from "./pages/About/About";
-import Contact from "./pages/Contact/Contact";
+import About from "./pages/about/About";
+import Contact from "./pages/contact";
 import Signin from "./pages/auth/Signin";
 import Register from "./pages/auth/Register";
-import Home from "./pages/Home/Home";
-import Product from "./pages/Product/Product";
-import Products from "./pages/Products/Products";
+import Home from "./pages/home";
+import Product from "./pages/product";
+import Products from "./pages/products";
 import { Layout } from "./layouts/PublicLayout";
-import { Profile } from "./pages/users/Profile";
+import { Profile } from "./pages/users/profile";
 export const router = createBrowserRouter([
     {
         path: "/",
